app-trent/requirements.txt
pyprimes==0.2.2a0
pycrypto
requests
gunicorn
Flask
WTForms
Flask-WTF
jwcrypto